refactor:cleaning up un-used files
