--- src/app/privacy/page.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function page() {
-  return <div>Privecy policy</div>;
-}
