+++ src/components/modals/CreateBookingsFormDialog.tsx
@@ -7,7 +7,7 @@ import React, {
   useEffect,
   useCallback,
 } from "react";
-import { auth } from "../../firebase";
+import { auth } from "../../../firebase";
 import {
   Dialog,
   DialogContent,
